Add local directory "--buildpack" to pack build
This shortens the dev cycle for buildpack authors

[https://github.com/buildpack/pack/issues/54]

NOTE: We chose to copy the buildpack in to "detect" and "build" rather
than copy to workspace and reuse it, since changing the buildpacks
directory causes problems with the "/buildpacks" directory structure.
Specifically the fact that "latest" is often a symlink to a version.